Update PriceCalculator.sol
Modify line 153 to include option amount in utilization --> multiplier is now based on the future lockedAmount, not current; this ensures big orders pay their fair share

Declare "percentAbove" variable which calculates the percent of the option amount that lies above 40% utilization mark; "if" statement forces it to be between 0 and 1.  Previously I had included "else if (percentAbove < 0) percentAbove = 0;" but it's redundant because line 154 ensures that can never happen already

Include the percentAbove factor in the iv multiplication line --> any option amount that would bring utilization above 40% is only multiplied by the net amount it does bring over 40%, and not the entire amount of the option.

See conversation between abipup and southseacompany on the Hegic Discord server, v888-1-discussion channel on 3 March 2021 for more details.